move column name parser to separate py file
